Refactor edit page
